File restructure to be cleaner and more standard
