+++ README.md
@@ -1,2 +1,14 @@
 # covenant-finance
 A program for the purpose of yearly advancement and tracking of covenant finances and advancement.
+
+TODO:
+  * Example usage
+  * Add "Advance year" function
+  * Front end tests
+
+BUGS TO FIX:
+  * Login with non existant user throws error
+  * Fail covenant validation throws error page
+  * Fail lab validation (unique name) throws error page
+  * Covenant names are currently globally unique, need to be unique per user
+  * User's covenants do not populate until new login (create covenant redirect missing data)
