+++ tests/unit/covenant_tests.py
@@ -3,10 +3,10 @@
 
 import pytest
 
-from covenfolk import Covenfolk
-from covenant import Covenant
-from armory import Armory
-from sample_covenants import semita
+from src.covenfolk import Covenfolk
+from src.covenant import Covenant
+from src.armory import Armory
+from src.sample_covenants import semita
 
 covenfolken = {
     'magi' : 6,
@@ -225,7 +225,7 @@ class DescribeCovenant:
 class DescribeSaveCovenant:
     @staticmethod
     def it_saves_covenant():
-        from covenant import save_covenant
+        from src.covenant import save_covenant
         cov = Covenant()
         cov.inflation = 100
         cov.covenfolken.add_covenfolk("Barg", "magi")
@@ -242,7 +242,7 @@ class DescribeSaveCovenant:
 class DescribeLoadCovenant:
     @staticmethod
     def it_loads_covenant():
-        from covenant import save_covenant, load_covenant_from_file
+        from src.covenant import save_covenant, load_covenant_from_file
         cov = Covenant()
         cov.inflation = 100
         cov.covenfolken.add_covenfolk("Barg", "magi")
